extract mempool as small CMake module
