+++ CMakeLists.txt
@@ -238,6 +238,7 @@ add_subdirectory(src/inverted_index)
 add_subdirectory(src/wildcard)
 add_subdirectory(src/util/arr)
 add_subdirectory(src/util/hash)
+add_subdirectory(src/util/mempool)
 add_subdirectory(src/coord)
 
 #----------------------------------------------------------------------------------------------
@@ -282,6 +283,7 @@ add_library(rscore OBJECT ${SOURCES})
 set(FINAL_OBJECTS
     $<TARGET_OBJECTS:arr>
     $<TARGET_OBJECTS:inverted_index>
+    $<TARGET_OBJECTS:mempool>
     $<TARGET_OBJECTS:wildcard>
     $<TARGET_OBJECTS:rscore>
     $<TARGET_OBJECTS:rmutil>
@@ -313,6 +315,7 @@ target_link_libraries(redisearch
         redisearch-coord
         arr
         inverted_index
+        mempool
         wildcard
         uv_a
         ${BINDIR}/redisearch_rs/libredisearch_rs.a
